requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
Flask==2.3.2
fonttools==4.40.0
fsspec==2023.6.0
huggingface-hub==0.16.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
safetensors==0.3.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
six==1.16.0
sympy==1.12
torch==2.0.1+cu118
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu118
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
